Pull publish behavior into a hook
